backend/python-worker/requirements.txt
grpcio>=1.68
grpcio-health-checking>=1.68
grpcio-tools>=1.68
numpy>=2.0
soundfile>=0.12
funasr>=1.2
modelscope>=1.22
torch>=2.5
torchaudio>=2.5
pydantic>=2.0
